requirements.txt
elasticsearch
python-dateutil